requirements.txt
fastapi
uvicorn
python-dotenv
torch
boto3
pydantic
beautifulsoup4
sqlalchemy